pyproject.toml
[project]
name = "django-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-unfold>=0.72.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "pillow>=11.0.0",
    "django-filter>=24.3",
]
